requirements.txt
flask
flask_restfull
opencv-python
imutils